Import `typing` if annotation has been added and it is not imported yet & add `--check` to CLI & return 1 if changes have been made (#11)

+++ auto_typing_final/main.py
@@ -1,19 +1,31 @@
 import argparse
+import sys
+from difflib import ndiff
 from typing import TextIO, cast
 
 from auto_typing_final.transform import transform_file_content
 
 
-def main() -> None:  # pragma: no cover
+def main() -> int:
     parser = argparse.ArgumentParser()
     parser.add_argument("files", type=argparse.FileType("r+"), nargs="*")
+    parser.add_argument("--check", action="store_true")
+    args = parser.parse_args()
 
-    for file in cast(list[TextIO], parser.parse_args().files):
+    has_changes = False
+
+    for file in cast(list[TextIO], args.files):
         data = file.read()
-        file.seek(0)
-        file.write(transform_file_content(data))
-        file.truncate()
+        transformed_content = transform_file_content(data)
 
+        if args.check:
+            sys.stdout.writelines(ndiff(data.splitlines(keepends=True), transformed_content.splitlines(keepends=True)))
+        else:
+            file.seek(0)
+            file.write(transformed_content)
+            file.truncate()
 
-if __name__ == "__main__":  # pragma: no cover
-    main()
+        if data != transformed_content:
+            has_changes = True
+
+    return has_changes
